+++ trunk/controlador/Compras/GestorGenerarOrdenCompra.java
@@ -208,7 +208,7 @@ public GestorGenerarOrdenCompra(pantallaGenerarOrdenCompra pantalla)
           {
               NTupla nt=new NTupla();
               nt.setId(listaPrecios.get(i).getId());
-              nt.setNombre(Integer.toString(listaPrecios.get(i).getCantidad()));
+              nt.setNombre(Double.toString(listaPrecios.get(i).getCantidad()));
               String[] datos=new String[3];
 
               datos[0]=Double.toString(listaPrecios.get(i).getPrecio());
